refactor(perception): split stereo_point_cloud into package with vio.py + utils.py
