+++ dimos/perception/stereo_point_cloud/module.py
@@ -21,8 +21,6 @@ from typing import Any
 
 import numpy as np
 from reactivex.disposable import Disposable
-from scipy.ndimage import sobel
-from scipy.spatial import cKDTree
 
 from dimos.core.core import rpc
 from dimos.core.module import Module, ModuleConfig
@@ -30,205 +28,18 @@ from dimos.core.stream import In, Out
 from dimos.msgs.sensor_msgs.CameraInfo import CameraInfo
 from dimos.msgs.sensor_msgs.Image import Image
 from dimos.msgs.sensor_msgs.PointCloud2 import PointCloud2
+from dimos.perception.stereo_point_cloud.utils import (
+    _FloorCalibrator,
+    _R_OPT_TO_LINK,
+    _gradient_mask,
+    _pack,
+    _raycast_free_keys,
+)
+from dimos.perception.stereo_point_cloud.vio import MadgwickFilter, PointCloudOdometry
 from dimos.utils.logging_config import setup_logger
 
 logger = setup_logger()
 
-
-# ── Constants ─────────────────────────────────────────────────────────────────
-
-_VOFF  = np.int64(100_000)
-_VMASK = np.int64(0x3FFFF)
-
-# Optical frame (X=right, Y=down, Z=depth) → camera_link (X=fwd, Y=left, Z=up)
-_R_OPT_TO_LINK = np.array([[0, 0, 1], [-1, 0, 0], [0, -1, 0]], dtype=np.float32)
-
-
-# ── Helpers ───────────────────────────────────────────────────────────────────
-
-def _pack(vkeys: np.ndarray) -> np.ndarray:
-    v = (vkeys.astype(np.int64) + _VOFF) & _VMASK
-    return (v[:, 0] << np.int64(36)) | (v[:, 1] << np.int64(18)) | v[:, 2]
-
-
-def _gradient_mask(depth: np.ndarray, threshold: float) -> np.ndarray:
-    valid   = np.isfinite(depth)
-    depth_f = np.where(valid, depth, 0.0).astype(np.float64)
-    grad    = np.hypot(sobel(depth_f, axis=1), sobel(depth_f, axis=0))
-    return valid & (grad < threshold)
-
-
-def _raycast_free_keys(surface_pts: np.ndarray, vox_size: float, n_rays: int = 400) -> np.ndarray:
-    """Free-space voxel keys on rays from camera origin (rotation-only frame) to surface_pts."""
-    origin = np.zeros(3, dtype=np.float32)
-    n = min(len(surface_pts), n_rays)
-    if n == 0:
-        return np.empty(0, dtype=np.int64)
-    idx   = np.random.choice(len(surface_pts), n, replace=False) if len(surface_pts) > n else np.arange(n)
-    pts   = surface_pts[idx]
-    vecs  = pts - origin
-    dists = np.linalg.norm(vecs, axis=1)
-    ok    = dists > vox_size * 2
-    vecs, dists = vecs[ok], dists[ok]
-    if not len(vecs):
-        return np.empty(0, dtype=np.int64)
-    dirs      = vecs / dists[:, None]
-    max_steps = int(np.ceil(dists.max() / vox_size))
-    t_vals    = np.arange(max_steps, dtype=np.float32) * vox_size
-    t_end     = (dists - vox_size * 1.5)[:, None]
-    valid_m   = (t_vals[None, :] >= 0) & (t_vals[None, :] < t_end)
-    ray_pts   = origin + dirs[:, None, :] * t_vals[None, :, None]
-    vk_flat   = np.floor(ray_pts.reshape(-1, 3) / vox_size).astype(np.int32)
-    return np.unique(_pack(vk_flat)[valid_m.ravel()])
-
-
-# ── Floor calibration ─────────────────────────────────────────────────────────
-
-class _FloorCalibrator:
-    """Histogram-based one-shot floor height estimator in camera_link frame (Z=up)."""
-
-    SKIP_FRAMES  = 30
-    CALIB_FRAMES = 30
-    BIN_M        = 0.02
-
-    def __init__(self) -> None:
-        self._frame   = 0
-        self._samples: list[float] = []
-        self.floor_z:    float | None = None
-        self.cam_height: float | None = None
-
-    @property
-    def ready(self) -> bool:
-        return self.floor_z is not None
-
-    def update(self, xyz_cam: np.ndarray) -> None:
-        self._frame += 1
-        if self.ready or self._frame <= self.SKIP_FRAMES:
-            return
-        z    = xyz_cam[:, 2]
-        dist = np.linalg.norm(xyz_cam[:, :2], axis=1)
-        mask = (z < -0.30) & (dist < 4.0)
-        if mask.sum() < 200:
-            return
-        z_floor = z[mask]
-        lo, hi  = z_floor.min(), z_floor.max()
-        if hi - lo < self.BIN_M:
-            return
-        bins          = np.arange(lo, hi + self.BIN_M, self.BIN_M)
-        counts, edges = np.histogram(z_floor, bins=bins)
-        self._samples.append(float(edges[int(np.argmax(counts))] + self.BIN_M / 2))
-        if len(self._samples) >= self.CALIB_FRAMES:
-            self.floor_z    = float(np.median(self._samples))
-            self.cam_height = -self.floor_z
-            logger.info(f"StereoPointCloud: floor calibrated — camera {self.cam_height:.3f} m above floor")
-
-
-# ── Madgwick AHRS ─────────────────────────────────────────────────────────────
-
-class MadgwickFilter:
-    """Gyro + accel → orientation R (camera_link → world, Z-up). beta=0.033."""
-
-    def __init__(self, beta: float = 0.033) -> None:
-        self._q      = np.array([1., 0., 0., 0.], dtype=np.float64)
-        self._beta   = beta
-        self._t_prev: float | None = None
-
-    def update(self, gyro: np.ndarray, accel: np.ndarray, t: float) -> None:
-        if self._t_prev is None:
-            self._t_prev = t
-            return
-        dt = min(t - self._t_prev, 0.1)
-        self._t_prev = t
-        if dt <= 0:
-            return
-        q0, q1, q2, q3 = self._q
-        gx, gy, gz = gyro.astype(np.float64)
-        a_n = np.linalg.norm(accel)
-        if a_n > 0.5:
-            ax, ay, az = accel.astype(np.float64) / a_n
-            f1 = 2*(q1*q3 - q0*q2) - ax
-            f2 = 2*(q0*q1 + q2*q3) - ay
-            f3 = 2*(0.5 - q1**2 - q2**2) - az
-            J  = np.array([
-                [-2*q2,  2*q3, -2*q0, 2*q1],
-                [ 2*q1,  2*q0,  2*q3, 2*q2],
-                [    0, -4*q1, -4*q2,    0],
-            ])
-            grad = J.T @ np.array([f1, f2, f3])
-            gn   = np.linalg.norm(grad)
-            if gn > 1e-9:
-                grad /= gn
-        else:
-            grad = np.zeros(4)
-        q_dot = 0.5 * np.array([
-            -q1*gx - q2*gy - q3*gz,
-             q0*gx + q2*gz - q3*gy,
-             q0*gy - q1*gz + q3*gx,
-             q0*gz + q1*gy - q2*gx,
-        ]) - self._beta * grad
-        self._q = self._q + q_dot * dt
-        self._q /= np.linalg.norm(self._q) + 1e-12
-
-    @property
-    def R(self) -> np.ndarray:
-        q0, q1, q2, q3 = self._q
-        return np.array([
-            [1-2*(q2**2+q3**2), 2*(q1*q2-q0*q3),   2*(q1*q3+q0*q2)  ],
-            [2*(q1*q2+q0*q3),   1-2*(q1**2+q3**2), 2*(q2*q3-q0*q1)  ],
-            [2*(q1*q3-q0*q2),   2*(q2*q3+q0*q1),   1-2*(q1**2+q2**2)],
-        ], dtype=np.float32)
-
-
-# ── Point-cloud odometry ──────────────────────────────────────────────────────
-
-class PointCloudOdometry:
-    """Rotation-decoupled ICP for translation t. Takes R from Madgwick, converges in 3-4 iters."""
-
-    ITERS    = 4
-    MAX_DIST = 0.40
-    N_SRC    = 1_500
-    N_DST    = 8_000
-    N_STORE  = 10_000
-
-    def __init__(self) -> None:
-        self._t          = np.zeros(3, dtype=np.float32)
-        self._lock       = threading.Lock()
-        self._prev_world: np.ndarray | None = None
-
-    @property
-    def t(self) -> np.ndarray:
-        with self._lock:
-            return self._t.copy()
-
-    def update(self, xyz_cam: np.ndarray, R: np.ndarray) -> np.ndarray:
-        pts_ga = (xyz_cam @ R.T).astype(np.float32)
-        with self._lock:
-            t_est = self._t.copy()
-        if self._prev_world is None or len(pts_ga) < 50:
-            t_new = t_est
-        else:
-            n_src = min(self.N_SRC, len(pts_ga))
-            n_dst = min(self.N_DST, len(self._prev_world))
-            src   = pts_ga[np.random.choice(len(pts_ga), n_src, replace=False)]
-            dst   = self._prev_world[np.random.choice(len(self._prev_world), n_dst, replace=False)]
-            tree  = cKDTree(dst)
-            for _ in range(self.ITERS):
-                dists, idx = tree.query(src + t_est, k=1, workers=1)
-                mask = dists < self.MAX_DIST
-                if mask.sum() < 30:
-                    break
-                delta = dst[idx[mask]].mean(axis=0) - (src[mask] + t_est).mean(axis=0)
-                t_est = (t_est + 0.7 * delta).astype(np.float32)
-            t_new = t_est
-        n_st  = min(self.N_STORE, len(pts_ga))
-        idx_s = np.random.choice(len(pts_ga), n_st, replace=False)
-        self._prev_world = (pts_ga[idx_s] + t_new).astype(np.float32)
-        with self._lock:
-            self._t = t_new
-        return t_new
-
-
-# ── Module ────────────────────────────────────────────────────────────────────
 
 class Config(ModuleConfig):
     min_depth: float          = 0.1
